# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_pranavbh-9117_IMB_internal_quiz_dto.CreateQuestionRequest:
    properties:
      marks:
        minimum: 1
        type: integer
      options:
        items:
          $ref: '#/definitions/github_com_pranavbh-9117_IMB_internal_quiz_dto.OptionRequest'
        minItems: 2
        type: array
      text:
        type: string
    required:
    - marks
    - options
    - text
    type: object
  github_com_pranavbh-9117_IMB_internal_quiz_dto.CreateQuizRequest:
    properties:
      description:
        type: string
      duration_minutes:
        minimum: 1
        type: integer
      title:
        maxLength: 255
        type: string
    required:
    - duration_minutes
    - title
    type: object
  github_com_pranavbh-9117_IMB_internal_quiz_dto.OptionRequest:
    properties:
      is_correct:
        type: boolean
      text:
        maxLength: 255
        type: string
    required:
    - text
    type: object
  github_com_pranavbh-9117_IMB_internal_quiz_dto.OptionResponse:
    properties:
      id:
        type: string
      is_correct:
        description: May be hidden from students depending on endpoint
        type: boolean
      order_index:
        type: integer
      question_id:
        type: string
      text:
        type: string
    type: object
  github_com_pranavbh-9117_IMB_internal_quiz_dto.QuestionResponse:
    properties:
      id:
        type: string
      marks:
        type: integer
      options:
        items:
          $ref: '#/definitions/github_com_pranavbh-9117_IMB_internal_quiz_dto.OptionResponse'
        type: array
      order_index:
        type: integer
      quiz_id:
        type: string
      text:
        type: string
    type: object
  github_com_pranavbh-9117_IMB_internal_quiz_dto.QuizResponse:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      duration_minutes:
        type: integer
      id:
        type: string
      institution_id:
        type: string
      is_published:
        type: boolean
      questions:
        items:
          $ref: '#/definitions/github_com_pranavbh-9117_IMB_internal_quiz_dto.QuestionResponse'
        type: array
      title:
        type: string
      total_marks:
        type: integer
      updated_at:
        type: string
    type: object
  github_com_pranavbh-9117_IMB_internal_quiz_dto.UpdateQuizRequest:
    properties:
      description:
        type: string
      duration_minutes:
        minimum: 1
        type: integer
      title:
        maxLength: 255
        type: string
    required:
    - duration_minutes
    - title
    type: object
...
  github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse:
    properties:
      data: {}
      message:
        example: error description
        type: string
      success:
        example: false
        type: boolean
    type: object
  github_com_pranavbh-9117_IMB_pkg_response.SwaggerResponse-any:
    properties:
      data: {}
      message:
        example: operation successful
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  github_com_pranavbh-9117_IMB_pkg_response.SwaggerResponse-array_github_com_pranavbh-9117_IMB_internal_quiz_dto_QuizResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/github_com_pranavbh-9117_IMB_internal_quiz_dto.QuizResponse'
        type: array
      message:
        example: operation successful
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  github_com_pranavbh-9117_IMB_pkg_response.SwaggerResponse-github_com_pranavbh-9117_IMB_internal_quiz_dto_QuestionResponse:
    properties:
      data:
        $ref: '#/definitions/github_com_pranavbh-9117_IMB_internal_quiz_dto.QuestionResponse'
      message:
        example: operation successful
        type: string
      success:
        example: true
        type: boolean
    type: object
  github_com_pranavbh-9117_IMB_pkg_response.SwaggerResponse-github_com_pranavbh-9117_IMB_internal_quiz_dto_QuizResponse:
    properties:
      data:
        $ref: '#/definitions/github_com_pranavbh-9117_IMB_internal_quiz_dto.QuizResponse'
      message:
        example: operation successful
        type: string
      success:
        example: true
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API documentation for the Institute Management Platform.
  title: Institute Management Backend API
  version: "1.0"
paths:
...
  /quizzes:
    get:
      description: Retrieves a list of quizzes. Faculty see their own quizzes. Students
        see published quizzes for their institution.
      produces:
      - application/json
      responses:
        "200":
          description: Quizzes Retrieved
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerResponse-array_github_com_pranavbh-9117_IMB_internal_quiz_dto_QuizResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: List Quizzes
      tags:
      - Quiz
    post:
      consumes:
      - application/json
      description: Creates a new quiz. Faculty only.
      parameters:
      - description: Quiz Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_pranavbh-9117_IMB_internal_quiz_dto.CreateQuizRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Quiz Created
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerResponse-github_com_pranavbh-9117_IMB_internal_quiz_dto_QuizResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: Create Quiz
      tags:
      - Quiz
  /quizzes/{id}:
    delete:
      description: Deletes an existing quiz. Faculty only. Quizzes with attempts cannot
        be deleted.
      parameters:
      - description: Quiz ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Quiz Deleted
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerResponse-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete Quiz
      tags:
      - Quiz
    get:
      description: Retrieves a quiz by its ID. Faculty can view their own quizzes.
        Students can view published quizzes for their institution.
      parameters:
      - description: Quiz ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Quiz Retrieved
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerResponse-github_com_pranavbh-9117_IMB_internal_quiz_dto_QuizResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Quiz
      tags:
      - Quiz
    put:
      consumes:
      - application/json
      description: Updates an existing draft quiz metadata. Faculty only. Published
        quizzes cannot be modified.
      parameters:
      - description: Quiz ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_pranavbh-9117_IMB_internal_quiz_dto.UpdateQuizRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Quiz Updated
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerResponse-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: Update Quiz
      tags:
      - Quiz
...
  /quizzes/{id}/publish:
    put:
      description: Publishes a quiz, making it available to students. Faculty only.
      parameters:
      - description: Quiz ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Quiz Published
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerResponse-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: Publish Quiz
      tags:
      - Quiz
  /quizzes/{id}/questions:
    post:
      consumes:
      - application/json
      description: Adds a new question with options to a draft quiz. Faculty only.
      parameters:
      - description: Quiz ID
        in: path
        name: id
        required: true
        type: string
      - description: Question Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_pranavbh-9117_IMB_internal_quiz_dto.CreateQuestionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Question Created
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerResponse-github_com_pranavbh-9117_IMB_internal_quiz_dto_QuestionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: Create Question
      tags:
      - Quiz
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"